backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
openai==1.3.5
pydantic==2.4.2
pillow==10.1.0
serpapi==0.1.0
